fix(): nats bridge refactor
