+++ clients/openframe-chat/src-tauri/src/token_watcher.rs
@@ -2,6 +2,7 @@ use std::fs;
 use std::path::{Path, PathBuf};
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, SystemTime};
+use crate::nats_bridge::mask_token;
 use crate::token_decryption_service::TokenDecryptionService;
 use tauri::{AppHandle, Emitter};
 use serde::Serialize;
@@ -15,19 +16,17 @@ struct TokenUpdateEvent {
 ///
 /// Created empty and `enable`d once the daemon-written config is available, so
 /// it survives main's late/recovery config path. All clones share the same
-/// source + cache, so enabling one (from `apply_config`) makes the token
-/// visible to every holder (the NATS bridge, `get_token`, the watcher thread).
+/// source, so enabling one (from `apply_config`) makes the token visible to
+/// every holder (the NATS bridge, `get_token`, the watcher thread).
 ///
-/// Two read paths:
-///   - `read_fresh` decrypts the file on demand and updates the cache. Used on
-///     the NATS reconnect path (`auth_url_callback`) and by `get_token` so a
-///     (re)connect or a frontend refresh always sees the newest token the
-///     daemon has written, with zero dependency on poll timing.
-///   - `current` returns the last cached value without touching disk.
+/// `read_fresh` decrypts the file on demand. Used on the NATS reconnect path
+/// (`auth_url_callback`) and by `get_token` so a (re)connect or a frontend
+/// refresh always sees the newest token the daemon has written, with zero
+/// dependency on poll timing. `TokenWatcher` polls the same file and pushes
+/// `token-update` events to the WebView when it rotates.
 #[derive(Clone)]
 pub struct TokenSource {
     inner: Arc<Mutex<Option<TokenSourceInner>>>,
-    cached: Arc<Mutex<Option<String>>>,
 }
 
 struct TokenSourceInner {
@@ -40,7 +39,6 @@ impl TokenSource {
     pub fn new() -> Self {
         Self {
             inner: Arc::new(Mutex::new(None)),
-            cached: Arc::new(Mutex::new(None)),
         }
     }
 
@@ -59,19 +57,11 @@ impl TokenSource {
         true
     }
 
-    /// Reads + decrypts the token file now, refreshes the cache, returns it.
+    /// Reads + decrypts the token file now.
     pub fn read_fresh(&self) -> Option<String> {
-        let token = {
-            let guard = self.inner.lock().unwrap();
-            let inner = guard.as_ref()?;
-            read_and_decrypt(&inner.path, &inner.decryptor)
-        };
-        *self.cached.lock().unwrap() = token.clone();
-        token
-    }
-
-    pub fn current(&self) -> Option<String> {
-        self.cached.lock().unwrap().clone()
+        let guard = self.inner.lock().unwrap();
+        let inner = guard.as_ref()?;
+        read_and_decrypt(&inner.path, &inner.decryptor)
     }
 
     fn path(&self) -> Option<PathBuf> {
@@ -107,6 +97,11 @@ impl TokenWatcher {
 
         std::thread::spawn(move || {
             let mut last_mtime: Option<SystemTime> = None;
+            // Compare against the last token *emitted to the WebView*, not a
+            // shared cache — `read_fresh` from the NATS auth callback or
+            // `get_token` can read the file first, which would make a
+            // cache-based comparison swallow the rotation event.
+            let mut last_emitted: Option<String> = None;
 
             loop {
                 let mtime = fs::metadata(&path).and_then(|m| m.modified()).ok();
@@ -114,16 +109,22 @@ impl TokenWatcher {
                 if mtime.is_none() || mtime != last_mtime {
                     last_mtime = mtime;
 
-                    let prev = source.current();
                     let new = source.read_fresh();
-                    if prev != new {
+                    if new != last_emitted {
                         if let Some(token) = &new {
-                            match prev {
-                                None => log::info!("token watcher: first token received"),
-                                Some(_) => log::info!("token watcher: token refreshed"),
+                            match last_emitted {
+                                None => log::info!(
+                                    "token watcher: first token received ({})",
+                                    mask_token(token)
+                                ),
+                                Some(_) => log::info!(
+                                    "token watcher: token refreshed ({})",
+                                    mask_token(token)
+                                ),
                             }
                             emit_token_to_frontend(&app_handle, token);
                         }
+                        last_emitted = new;
                     }
                 }
 
@@ -137,7 +138,9 @@ fn emit_token_to_frontend(app_handle: &AppHandle, token: &str) {
     let event = TokenUpdateEvent {
         token: token.to_string(),
     };
-    match app_handle.emit("token-update", event) {
+    // emit_to: a broadcast `emit` reaches every event target, so a single JS
+    // `listen` would receive the event once per target (duplicates).
+    match app_handle.emit_to("main", "token-update", event) {
         Ok(_) => log::debug!("token watcher: token emitted to frontend"),
         Err(e) => log::error!("token watcher: failed to emit token-update event: {}", e),
     }
