+++ clients/openframe-chat/src/config/features.ts
@@ -4,6 +4,7 @@ export const FEATURE_FLAG_NAMES = [
   'ai-streaming-jetstream',
   'ticket-statuses',
   'customer-ai-assistant-settings',
+  'notifications',
 ] as const;
 
 export type FeatureFlagName = (typeof FEATURE_FLAG_NAMES)[number];
@@ -14,6 +15,7 @@ export const DEFAULT_FEATURE_FLAGS: Record<FeatureFlagName, boolean> = {
   'ai-streaming-jetstream': false,
   'ticket-statuses': false,
   'customer-ai-assistant-settings': false,
+  notifications: false,
 };
 
 export type FeatureFlags = Record<FeatureFlagName, boolean>;
